Add Barcode Scanner Settings and Nutrition Editor Features
- Implemented BarcodeScanner.SettingsView for configuring AI Nutrition Scanner settings, including model download and management.
- Created NutritionEditorSheet for editing nutritional information of scanned products.
- Added ProductDetailsView and ScannedProductRow for displaying product details and managing scanned products.
- Introduced ScannerPreviewView for real-time barcode scanning with camera integration.
- Updated ServicesView to include navigation to Barcode Scanner settings.
- Enhanced routing to support new barcode scanner settings screen.
