+++ Trio/Sources/Modules/BarcodeAi/BarcodeAiDataFlow.swift
@@ -1,5 +1,7 @@
+/// BarcodeScanner module for scanning barcodes and nutrition labels
 enum BarcodeScanner {
     enum Config {}
 }
 
+/// Provider protocol for BarcodeScanner module
 protocol BarcodeScannerProvider: Provider {}
