+++ Trio/Sources/Modules/BarcodeAi/BarcodeScannerSettings/BarcodeScannerSettingsDataFlow.swift
@@ -1,0 +1,5 @@
+enum BarcodeScannerSettings {
+    enum Config {}
+}
+
+protocol BarcodeScannerSettingsProvider: Provider {}
