+++ Trio/Sources/Modules/BarcodeAi/BarcodeScannerSettings/BarcodeScannerSettingsProvider.swift
@@ -1,0 +1,3 @@
+extension BarcodeScannerSettings {
+    final class Provider: BaseProvider, BarcodeScannerSettingsProvider {}
+}
