+++ Trio/Sources/Modules/BarcodeAi/View/ProductViews.swift
@@ -1,0 +1,310 @@
+import SwiftUI
+
+// MARK: - Product Details View
+
+extension BarcodeScanner {
+    struct ProductDetailsView: View {
+        let product: OpenFoodFactsProduct
+
+        var body: some View {
+            VStack(alignment: .leading, spacing: 16) {
+                HStack(alignment: .top, spacing: 16) {
+                    if let imageURL = product.imageURL {
+                        AsyncImage(url: imageURL) { phase in
+                            switch phase {
+                            case let .success(image):
+                                image
+                                    .resizable()
+                                    .scaledToFill()
+                            case .failure:
+                                placeholder
+                            default:
+                                ProgressView()
+                            }
+                        }
+                        .frame(width: 88, height: 88)
+                        .clipShape(RoundedRectangle(cornerRadius: 16))
+                    } else {
+                        placeholder
+                            .frame(width: 88, height: 88)
+                    }
+
+                    VStack(alignment: .leading, spacing: 6) {
+                        Text(product.name)
+                            .font(.headline)
+                        if let brand = product.brand {
+                            Text(brand)
+                                .font(.subheadline)
+                                .foregroundStyle(.secondary)
+                        }
+                        if let quantity = product.quantity {
+                            Text(quantity)
+                                .font(.footnote)
+                                .foregroundStyle(.secondary)
+                        }
+                        if let serving = product.servingSize {
+                            Text("Serving: \(serving)")
+                                .font(.footnote)
+                                .foregroundStyle(.secondary)
+                        }
+                    }
+                    Spacer()
+                }
+
+                NutrimentGrid(nutriments: product.nutriments)
+
+                if let ingredients = product.ingredients {
+                    VStack(alignment: .leading, spacing: 4) {
+                        Text("Ingredients")
+                            .font(.subheadline.weight(.semibold))
+                        Text(ingredients)
+                            .font(.footnote)
+                            .foregroundStyle(.secondary)
+                    }
+                }
+            }
+            .padding()
+            .background(.ultraThinMaterial)
+            .clipShape(RoundedRectangle(cornerRadius: 20))
+        }
+
+        private var placeholder: some View {
+            RoundedRectangle(cornerRadius: 16)
+                .fill(Color.secondary.opacity(0.2))
+                .overlay(
+                    Image(systemName: "photo")
+                        .foregroundStyle(.secondary)
+                )
+        }
+    }
+}
+
+// MARK: - Scanned Product Row
+
+extension BarcodeScanner {
+    struct ScannedProductRow: View {
+        let item: ScannedProductItem
+        var state: StateModel
+
+        @State private var amountText: String = ""
+        @State private var isMlInput: Bool = false
+        @State private var showQuickSelector: Bool = false
+        @FocusState private var isTextFieldFocused: Bool
+
+        var body: some View {
+            VStack(alignment: .leading, spacing: 12) {
+                HStack(alignment: .top, spacing: 12) {
+                    if let imageURL = item.product.imageURL {
+                        AsyncImage(url: imageURL) { phase in
+                            switch phase {
+                            case let .success(image):
+                                image
+                                    .resizable()
+                                    .scaledToFill()
+                            case .failure:
+                                placeholder
+                            default:
+                                ProgressView()
+                            }
+                        }
+                        .frame(width: 60, height: 60)
+                        .clipShape(RoundedRectangle(cornerRadius: 12))
+                    } else {
+                        placeholder
+                            .frame(width: 60, height: 60)
+                    }
+
+                    VStack(alignment: .leading, spacing: 4) {
+                        Text(item.product.name)
+                            .font(.subheadline.weight(.semibold))
+                            .lineLimit(2)
+                        if let brand = item.product.brand {
+                            Text(brand)
+                                .font(.caption)
+                                .foregroundStyle(.secondary)
+                        }
+                    }
+
+                    Spacer()
+
+                    Button(role: .destructive) {
+                        state.removeScannedProduct(item)
+                    } label: {
+                        Image(systemName: "trash")
+                            .foregroundStyle(.red)
+                    }
+                }
+
+                HStack(spacing: 12) {
+                    TextField(
+                        String(localized: "Amount"),
+                        text: $amountText
+                    )
+                    .keyboardType(.decimalPad)
+                    .textFieldStyle(.roundedBorder)
+                    .focused($isTextFieldFocused)
+                    .onChange(of: amountText) { _, newValue in
+                        if let amount = Double(newValue.replacingOccurrences(of: ",", with: ".")) {
+                            state.updateScannedProductAmount(item, amount: amount, isMlInput: isMlInput)
+                        }
+                    }
+
+                    Picker("", selection: $isMlInput) {
+                        Text("g").tag(false)
+                        Text("ml").tag(true)
+                    }
+                    .pickerStyle(.segmented)
+                    .frame(width: 100)
+                    .onChange(of: isMlInput) { _, newValue in
+                        if let amount = Double(amountText.replacingOccurrences(of: ",", with: ".")) {
+                            state.updateScannedProductAmount(item, amount: amount, isMlInput: newValue)
+                        }
+                    }
+                }
+
+                if showQuickSelector {
+                    VStack(spacing: 8) {
+                        HStack(spacing: 8) {
+                            ForEach(1 ... 5, id: \.self) { multiplier in
+                                Button {
+                                    if let servingQuantity = item.product.servingQuantity, servingQuantity > 0 {
+                                        updateAmount(servingQuantity * Double(multiplier))
+                                    } else {
+                                        updateAmount(Double(multiplier) * 100)
+                                    }
+                                    showQuickSelector = false
+                                } label: {
+                                    Text("\(multiplier)x")
+                                        .font(.caption.weight(.medium))
+                                        .frame(maxWidth: .infinity)
+                                }
+                                .buttonStyle(.bordered)
+                            }
+                        }
+                        HStack(spacing: 8) {
+                            ForEach(6 ... 10, id: \.self) { multiplier in
+                                Button {
+                                    if let servingQuantity = item.product.servingQuantity, servingQuantity > 0 {
+                                        updateAmount(servingQuantity * Double(multiplier))
+                                    } else {
+                                        updateAmount(Double(multiplier) * 100)
+                                    }
+                                    showQuickSelector = false
+                                } label: {
+                                    Text("\(multiplier)x")
+                                        .font(.caption.weight(.medium))
+                                        .frame(maxWidth: .infinity)
+                                }
+                                .buttonStyle(.bordered)
+                            }
+                        }
+                    }
+                    .transition(.opacity.combined(with: .move(edge: .top)))
+                }
+            }
+            .padding()
+            .background(.ultraThinMaterial)
+            .clipShape(RoundedRectangle(cornerRadius: 16))
+            .onTapGesture {
+                withAnimation(.easeInOut(duration: 0.2)) {
+                    showQuickSelector.toggle()
+                }
+            }
+            .onAppear {
+                updateFromItem()
+            }
+            .onChange(of: item.amount) { _, _ in
+                updateFromItem()
+            }
+            .onChange(of: item.isMlInput) { _, _ in
+                updateFromItem()
+            }
+            .onChange(of: isTextFieldFocused) { _, newValue in
+                // Pause scanner and hide scanner view when numpad is opened
+                state.isScanning = false
+                state.isKeyboardVisible = newValue
+            }
+        }
+
+        private func updateFromItem() {
+            if item.amount > 0 {
+                if item.amount.truncatingRemainder(dividingBy: 1) == 0 {
+                    amountText = String(format: "%.0f", item.amount)
+                } else {
+                    amountText = String(format: "%.1f", item.amount)
+                }
+            } else {
+                amountText = ""
+            }
+            isMlInput = item.isMlInput
+        }
+
+        private func updateAmount(_ amount: Double) {
+            if amount.truncatingRemainder(dividingBy: 1) == 0 {
+                amountText = String(format: "%.0f", amount)
+            } else {
+                amountText = String(format: "%.1f", amount)
+            }
+            state.updateScannedProductAmount(item, amount: amount, isMlInput: isMlInput)
+        }
+
+        private var placeholder: some View {
+            RoundedRectangle(cornerRadius: 12)
+                .fill(Color.secondary.opacity(0.2))
+                .overlay(
+                    Image(systemName: "photo")
+                        .foregroundStyle(.secondary)
+                        .font(.caption)
+                )
+        }
+    }
+}
+
+// MARK: - Nutriment Grid
+
+extension BarcodeScanner {
+    struct NutrimentGrid: View {
+        let nutriments: OpenFoodFactsProduct.Nutriments
+
+        var body: some View {
+            VStack(alignment: .leading, spacing: 8) {
+                Text("Per 100\(nutriments.basis == .per100ml ? "ml" : "g")")
+                    .font(.subheadline.weight(.semibold))
+
+                LazyVGrid(columns: Array(repeating: GridItem(.flexible(), spacing: 12), count: 2), spacing: 12) {
+                    NutrimentTile(title: String(localized: "Energy (kcal)"), value: nutriments.energyKcalPer100g, unit: "kcal")
+                    NutrimentTile(title: String(localized: "Carbs"), value: nutriments.carbohydratesPer100g, unit: "g")
+                    NutrimentTile(title: String(localized: "Sugars"), value: nutriments.sugarsPer100g, unit: "g")
+                    NutrimentTile(title: String(localized: "Fat"), value: nutriments.fatPer100g, unit: "g")
+                    NutrimentTile(title: String(localized: "Protein"), value: nutriments.proteinPer100g, unit: "g")
+                    NutrimentTile(title: String(localized: "Fiber"), value: nutriments.fiberPer100g, unit: "g")
+                }
+            }
+        }
+    }
+
+    struct NutrimentTile: View {
+        let title: String
+        let value: Double?
+        let unit: String
+
+        var body: some View {
+            VStack(alignment: .leading, spacing: 6) {
+                Text(title)
+                    .font(.caption)
+                    .foregroundStyle(.secondary)
+                Text(formattedValue)
+                    .font(.headline)
+            }
+            .padding()
+            .frame(maxWidth: .infinity, alignment: .leading)
+            .background(Color.secondary.opacity(0.08))
+            .clipShape(RoundedRectangle(cornerRadius: 12))
+        }
+
+        private var formattedValue: String {
+            guard let value else { return "—" }
+            return "\(String(format: "%.1f", value)) \(unit)"
+        }
+    }
+}
